+++ Program/.idea/misc.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="Black">
+    <option name="sdkName" value="Python 3.7 (while loop.py)" />
+  </component>
+  <component name="ProjectRootManager" version="2" project-jdk-name="Python 3.7 (while loop.py)" project-jdk-type="Python SDK" />
+</project>
