+++ Program/db.py
@@ -1,4 +1,7 @@
 import sqlite3
+import hashlib
+import os
+from security.encryption import encrypt, decrypt
 
 def create_connection():
     conn = sqlite3.connect('unique_meal.db')
@@ -8,7 +11,7 @@ def initialize_db():
     conn = create_connection()
     cursor = conn.cursor()
     # Create tables
-    cursor.execute('''CREATE TABLE IF NOT EXISTS menus (
+    cursor.execute('''CREATE TABLE IF NOT EXISTS users (
                         id INTEGER PRIMARY KEY AUTOINCREMENT,
                         username TEXT UNIQUE NOT NULL,
                         password_hash TEXT NOT NULL,
@@ -17,7 +20,7 @@ def initialize_db():
                         last_name TEXT,
                         registration_date TEXT)''')
     cursor.execute('''CREATE TABLE IF NOT EXISTS members (
-                        id INTEGER PRIMARY KEY AUTOINCREMENT,
+                        member_id INTEGER PRIMARY KEY,
                         first_name TEXT NOT NULL,
                         last_name TEXT NOT NULL,
                         age INTEGER NOT NULL,
@@ -26,8 +29,7 @@ def initialize_db():
                         address TEXT NOT NULL,
                         email TEXT NOT NULL,
                         phone TEXT NOT NULL,
-                        registration_date TEXT NOT NULL,
-                        member_id TEXT UNIQUE NOT NULL)''')
+                        registration_date TEXT NOT NULL)''')
     cursor.execute('''CREATE TABLE IF NOT EXISTS logs (
                         id INTEGER PRIMARY KEY AUTOINCREMENT,
                         date TEXT NOT NULL,
@@ -38,13 +40,64 @@ def initialize_db():
     conn.commit()
     conn.close()
 
+def add_user(username, password, role, first_name, last_name):
+    e_username = encrypt(username)
+    password_hash = hash_password(password)
+    e_role = encrypt(role)
+    e_first_name = encrypt(first_name)
+    e_last_name = encrypt(last_name)
+
+    conn = create_connection()
+    cursor = conn.cursor()
+
+    cursor.execute('INSERT OR IGNORE INTO users (username, password_hash, role, first_name, last_name, registration_date) VALUES (?, ?, ?, ?, ?, DATE())',
+                   (e_username, password_hash, e_role, e_first_name, e_last_name))
+    conn.commit()
+    conn.close()
+
+def get_role(username):
+    conn = create_connection()
+    cursor = conn.cursor()
+    cursor.execute('SELECT role FROM users WHERE username=?', (username,))
+    data = cursor.fetchone()[0]
+    conn.close()
+    print("encrypted role : " + data)
+    print("decrypted role : " + decrypt(data))
+    return decrypt(data)
+
+def search_member(query):
+    # Connect to SQLite database
+    conn = create_connection()
+    cursor = conn.cursor()
+
+    # Prepare the search query to search all fields
+    search_query = f"%{query}%"
+
+    # Query the database
+    cursor.execute('''
+       SELECT * FROM members
+       WHERE member_id LIKE ? OR
+             first_name LIKE ? OR
+             last_name LIKE ? OR
+             address LIKE ? OR
+             email LIKE ? OR
+             phone LIKE ?
+       ''', (search_query, search_query, search_query, search_query, search_query, search_query))
+
+    # Fetch all matching rows
+    results = cursor.fetchall()
+
+    # Close the connection
+    conn.close()
+
+    return results
 
 def display_all_info():
     conn = create_connection()
     cursor = conn.cursor()
 
     print("\nUsers:")
-    cursor.execute("SELECT * FROM menus")
+    cursor.execute("SELECT * FROM users")
     users = cursor.fetchall()
     for user in users:
         print(user)
@@ -72,3 +125,59 @@ def delete_table_users():
 
     display_all_info()
 
+def hash_password(password):
+    """
+    This function will return a hash of the provided password
+
+    When storing the hashed password we should also store salt to be able to verify later
+    advised format = salt:hashed_password
+
+    :param password: provided password by user (recommended to first use password_regex)
+    :return: hashed password
+    """
+
+    byte_password = password.encode()
+    salt = os.urandom(16)
+    iterations = 500_000
+    hash_name = 'sha256'
+    derived_key_length = None
+
+    result = hashlib.pbkdf2_hmac(hash_name, byte_password, salt, iterations, derived_key_length)
+
+    return f"{salt.hex()}:{result.hex()}"
+    #return hashlib.sha256(password.encode()).hexdigest()
+
+def verify_password(stored_password, input_password):
+    """
+    Verifies password provided by user with stored password in database
+
+    :param stored_password: hashed password stored in database
+    :param input_password: un-hashed password provided by user
+    :return: Boolean based on if the provided hashed password matches hashed password saved in database
+    """
+
+    salt = stored_password.split(':')[0]
+    hashed_password = stored_password.split(':')[1]
+
+    salt = bytes.fromhex(salt)
+    hashed_password = bytes.fromhex(hashed_password)
+
+    input_password = input_password.encode()
+    hash_name = 'sha256'
+    iterations = 500_000
+    derived_key_length = None
+
+    new_hashed_password = hashlib.pbkdf2_hmac(hash_name, input_password, salt, iterations, derived_key_length)
+
+    return new_hashed_password == hashed_password
+
+def authenticate(username, password):
+    conn = create_connection()
+    cursor = conn.cursor()
+    cursor.execute('SELECT password_hash FROM users WHERE username=?', (encrypt(username),))
+    data = cursor.fetchone() # Its not getting data - i think because we are encrypting data and it will encrypt differently again, you can test
+    conn.close()
+    return verify_password(data[0], password)
+
+def quick_auth(user_role, role):
+    return False if user_role != role else True
